chore: over-engineering cleanup round (ponytail audit of recent additions)
Audited everything added since e8a5f1cc with the laziness ladder;
six findings applied, net -69 LOC and one file fewer:

- Caps::preview_mcp collapsed into mcp_servers (identical for all five
  providers; preview attachment was the only distinct consumer)
- one crate-private MutexExt in term replaces three identical copies
- both UnknownReason taxonomies deleted (nothing read them — runtime
  matches Unknown with .. and stays deliberately silent)
- version_check three-file split folded to two, check() seams intact
- one-caller native_question_response wrapper inlined
- duplicate fixture export removed from the determinism snapshot test

Twenty further suspects were checked and rejected as earning their
keep (seeding safety rails, residency policy, search cache, export
policy header, smoke harness among them).
