+++ crates/agent/src/lib.rs
@@ -46,8 +46,8 @@ pub struct Caps {
     pub live_approval_mode_switch: bool,
     pub live_option_push: LiveOptionPush,
     /// Whether this adapter can attach tcode's HTTP MCP servers to a session.
+    /// Preview attachment rides the same capability as every other MCP server.
     pub mcp_servers: bool,
-    pub preview_mcp: bool,
     pub launch_args: bool,
     pub downgrade_approval_without_native_approvals: bool,
     pub option_descriptors: OptionDescriptors,
@@ -94,7 +94,6 @@ impl ProviderKind {
                 live_approval_mode_switch: true,
                 live_option_push: LiveOptionPush::None,
                 mcp_servers: true,
-                preview_mcp: true,
                 launch_args: true,
                 downgrade_approval_without_native_approvals: false,
                 option_descriptors: OptionDescriptors::Catalog,
@@ -110,7 +109,6 @@ impl ProviderKind {
                 live_approval_mode_switch: false,
                 live_option_push: LiveOptionPush::None,
                 mcp_servers: true,
-                preview_mcp: true,
                 launch_args: false,
                 downgrade_approval_without_native_approvals: false,
                 option_descriptors: OptionDescriptors::Catalog,
@@ -128,7 +126,6 @@ impl ProviderKind {
                 // ACP support is negotiated per agent; capable agents receive
                 // every tcode HTTP MCP registration in session/new or load.
                 mcp_servers: true,
-                preview_mcp: true,
                 launch_args: true,
                 downgrade_approval_without_native_approvals: false,
                 option_descriptors: OptionDescriptors::Wire,
@@ -144,7 +141,6 @@ impl ProviderKind {
                 live_approval_mode_switch: false,
                 live_option_push: LiveOptionPush::Only(&["reasoningEffort"]),
                 mcp_servers: false,
-                preview_mcp: false,
                 launch_args: true,
                 downgrade_approval_without_native_approvals: true,
                 option_descriptors: OptionDescriptors::Catalog,
@@ -160,7 +156,6 @@ impl ProviderKind {
                 live_approval_mode_switch: false,
                 live_option_push: LiveOptionPush::None,
                 mcp_servers: true,
-                preview_mcp: true,
                 launch_args: true,
                 downgrade_approval_without_native_approvals: false,
                 option_descriptors: OptionDescriptors::Catalog,
@@ -198,7 +193,6 @@ mod provider_caps_tests {
                 live_approval_mode_switch: true,
                 live_option_push: LiveOptionPush::None,
                 mcp_servers: true,
-                preview_mcp: true,
                 launch_args: true,
                 downgrade_approval_without_native_approvals: false,
                 option_descriptors: OptionDescriptors::Catalog,
@@ -217,7 +211,6 @@ mod provider_caps_tests {
                 live_approval_mode_switch: false,
                 live_option_push: LiveOptionPush::None,
                 mcp_servers: true,
-                preview_mcp: true,
                 launch_args: false,
                 downgrade_approval_without_native_approvals: false,
                 option_descriptors: OptionDescriptors::Catalog,
@@ -236,7 +229,6 @@ mod provider_caps_tests {
                 live_approval_mode_switch: false,
                 live_option_push: LiveOptionPush::All,
                 mcp_servers: true,
-                preview_mcp: true,
                 launch_args: true,
                 downgrade_approval_without_native_approvals: false,
                 option_descriptors: OptionDescriptors::Wire,
@@ -255,7 +247,6 @@ mod provider_caps_tests {
                 live_approval_mode_switch: false,
                 live_option_push: LiveOptionPush::Only(&["reasoningEffort"]),
                 mcp_servers: false,
-                preview_mcp: false,
                 launch_args: true,
                 downgrade_approval_without_native_approvals: true,
                 option_descriptors: OptionDescriptors::Catalog,
@@ -274,7 +265,6 @@ mod provider_caps_tests {
                 live_approval_mode_switch: false,
                 live_option_push: LiveOptionPush::None,
                 mcp_servers: true,
-                preview_mcp: true,
                 launch_args: true,
                 downgrade_approval_without_native_approvals: false,
                 option_descriptors: OptionDescriptors::Catalog,
