+++ crates/term/src/sync.rs
@@ -1,0 +1,11 @@
+use std::sync::{Mutex, MutexGuard};
+
+pub(crate) trait MutexExt<T> {
+    fn lock_recover(&self) -> MutexGuard<'_, T>;
+}
+
+impl<T> MutexExt<T> for Mutex<T> {
+    fn lock_recover(&self) -> MutexGuard<'_, T> {
+        self.lock().unwrap_or_else(|poisoned| poisoned.into_inner())
+    }
+}
